+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Title generation endpoint.
 - [frontend] - Dark mode.
 - [frontend] - Collapsible side bar.
+- [frontend] - Add confirmation box on thread deletion.
 
 ### Fixed
 - Migrate LLM Evaluation logic to scripts and add tests